presence validations
Note that validate_presence_of and validate_email use different
approaches (the former patches a function into self.validations, the
latter is such a function to be so patched during
self.__init__). Decisions will need to be made about which approach to
use and how to lift some of these up into serpentine_record.SqlObject
for general use.